Revert "Refactor monolithic script into modular Python package with enhanced features"
