--- .github/workflows/test.yml
@@ -1,44 +1,0 @@
-name: Tests
-
-on:
-  push:
-    branches: [ main, dev-refactor ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.10", "3.11", "3.12"]
-
-    steps:
-    - uses: actions/checkout@v4
-    
-    - name: Install uv
-      uses: astral-sh/setup-uv@v4
-      with:
-        version: "latest"
-        
-    - name: Set up Python ${{ matrix.python-version }}
-      run: uv python install ${{ matrix.python-version }}
-      
-    - name: Install dependencies
-      run: |
-        uv sync --dev
-        
-    - name: Run tests
-      run: |
-        uv run pytest tests/ -v --tb=short
-        
-    - name: Run tests with coverage
-      run: |
-        uv run pytest tests/ --cov=jwst_preprint_analyzer --cov-report=xml --cov-report=term
-        
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v3
-      if: matrix.python-version == '3.12'
-      with:
-        files: ./coverage.xml
-        fail_ci_if_error: false
