--- jwst_preprint_analyzer/__init__.py
@@ -1,3 +1,0 @@
-"""JWST Preprint DOI Analyzer package."""
-
-__version__ = "0.3.0"
