--- jwst_preprint_analyzer/analysis/__init__.py
@@ -1,1 +1,0 @@
-"""Analysis modules for JWST science and DOI detection."""
