--- jwst_preprint_analyzer/clients/__init__.py
@@ -1,1 +1,0 @@
-"""API client modules."""
