--- jwst_preprint_analyzer/processing/__init__.py
@@ -1,1 +1,0 @@
-"""Processing modules for paper download, conversion and text extraction."""
