--- jwst_preprint_analyzer/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utility modules for JWST preprint analyzer."""
