--- pyproject.toml
@@ -1,50 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "jwst-preprint-analyzer"
-version = "0.3.1"
-description = "Analyze arXiv preprints for JWST science content and DOI citations"
-readme = "README.md"
-authors = [
-    {name = "STScI"},
-]
-requires-python = ">=3.10"
-dependencies = [
-    "openai>=1.0.0",
-    "cohere>=5.0.0",
-    "nltk>=3.8",
-    "pydantic>=2.0.0",
-    "requests>=2.28.0",
-    "pdftext>=0.6.0",
-    "python-dotenv>=0.19.0",
-]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0",
-    "pytest-cov>=4.0",
-    "black>=22.0",
-    "ruff>=0.1.0",
-]
-
-[project.scripts]
-jwst-preprint-analyzer = "jwst_preprint_analyzer.__main__:main"
-
-[tool.setuptools.packages.find]
-include = ["jwst_preprint_analyzer*"]
-
-[tool.black]
-line-length = 100
-target-version = ['py38', 'py39', 'py310', 'py311']
-
-[tool.ruff]
-line-length = 100
-target-version = "py38"
-
-[dependency-groups]
-dev = [
-    "pytest>=8.4.0",
-    "pytest-cov>=6.1.1",
-]
