--- pytest.ini
@@ -1,18 +1,0 @@
-[tool:pytest]
-testpaths = tests
-python_files = test_*.py
-python_functions = test_*
-python_classes = Test*
-addopts = 
-    -v
-    --tb=short
-    --strict-markers
-    --disable-warnings
-filterwarnings =
-    ignore::DeprecationWarning
-    ignore::PendingDeprecationWarning
-markers =
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    integration: marks tests as integration tests
-    unit: marks tests as unit tests
-    skip_nltk: marks tests that require NLTK punkt tokenizer (skipped in CI)
